#2260 fix intel build
git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk@19208 b5f69653-2bf5-0310-8201-b0dafb8e1405

+++ mesh/src/utilities/DistributedBoxCollection.cpp
@@ -393,8 +393,8 @@ int DistributedBoxCollection<DIM>::LoadBalance(std::vector<int> localDistributio
 {
     MPI_Status status;
 
-    int proc_right = (PetscTools::AmTopMost()) ? MPI_PROC_NULL : PetscTools::GetMyRank() + 1;
-    int proc_left = (PetscTools::AmMaster()) ? MPI_PROC_NULL : PetscTools::GetMyRank() - 1;
+    int proc_right = (PetscTools::AmTopMost()) ? MPI_PROC_NULL : (int)PetscTools::GetMyRank() + 1;
+    int proc_left = (PetscTools::AmMaster()) ? MPI_PROC_NULL : (int)PetscTools::GetMyRank() - 1;
 
     // A variable that will return the new number of rows.
     int new_rows = localDistribution.size();
